Added tags to output when loading frames

+++ ase/tags.go
@@ -9,7 +9,7 @@ type Tag struct {
 	From                   uint16  `label:"From frame"`
 	To                     uint16  `label:"To frame"`
 	LoopAnimationDirection byte    `label:"Loop animation direction"`
-	Repeat                 byte    `label:"Repeat N Times"`
+	Repeat                 uint16  `label:"Repeat N Times"`
 	_                      [6]byte `label:"For future use"`
 	Color                  [3]byte `label:"RGB values of the tag color"`
 	_                      byte    `label:"Extra byte"`
